adexState context in separate file
