+++ src/app/PairsList.tsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useState } from "react";
-import { AdexStateContext } from "./page";
 import * as adex from "alphadex-sdk-js";
+import { AdexStateContext } from "./adex-state-context";
 
 export function PairsList() {
   const adexState = useContext(AdexStateContext);
